Move code to new utilization app
