+++ app/grandchallenge/workstations/models.py
@@ -2,7 +2,6 @@ import json
 import logging
 from datetime import datetime, timedelta
 from functools import cached_property
-from math import ceil
 from urllib.parse import unquote, urljoin
 
 from django.conf import settings
@@ -40,9 +39,6 @@ from grandchallenge.core.storage import (
     public_s3_storage,
 )
 from grandchallenge.core.validators import JSONValidator
-from grandchallenge.reader_studies.interactive_algorithms import (
-    InteractiveAlgorithmChoices,
-)
 from grandchallenge.subdomains.utils import reverse
 from grandchallenge.workstations.emails import send_new_feedback_email_to_staff
 
@@ -663,6 +659,8 @@ class Session(FieldChangeMixin, UUIDModel):
             )
 
         if self.has_changed("status") and self.status == self.STOPPED:
+            from grandchallenge.utilization.models import SessionCost
+
             SessionCost.objects.create(
                 session=self,
                 duration=now() - self.created,
@@ -733,92 +731,3 @@ class FeedbackGroupObjectPermission(GroupObjectPermissionBase):
     allowed_permissions = frozenset()
 
     content_object = models.ForeignKey(Feedback, on_delete=models.CASCADE)
-
-
-class SessionCost(UUIDModel):
-    session = models.OneToOneField(
-        Session,
-        related_name="session_cost",
-        null=True,
-        on_delete=models.SET_NULL,
-    )
-    duration = models.DurationField()
-    creator = models.ForeignKey(
-        settings.AUTH_USER_MODEL, null=True, on_delete=models.SET_NULL
-    )
-    reader_studies = models.ManyToManyField(
-        to="reader_studies.ReaderStudy",
-        through="SessionCostReaderStudy",
-        related_name="session_costs",
-        blank=True,
-        help_text="Reader studies accessed during session",
-    )
-    interactive_algorithms = models.JSONField(
-        blank=True,
-        default=list,
-        help_text=(
-            "The interactive algorithms for which hardware has been initialized during the session."
-        ),
-        validators=[
-            JSONValidator(
-                schema={
-                    "$schema": "http://json-schema.org/draft-07/schema",
-                    "type": "array",
-                    "items": {
-                        "enum": InteractiveAlgorithmChoices.values,
-                        "type": "string",
-                    },
-                    "uniqueItems": True,
-                }
-            )
-        ],
-    )
-
-    def save(self, *args, **kwargs) -> None:
-        from grandchallenge.reader_studies.models import Question
-
-        adding = self._state.adding
-
-        if adding:
-            reader_studies = self.session.reader_studies.all()
-            self.creator = self.session.creator
-            self.interactive_algorithms = list(
-                Question.objects.filter(reader_study__in=reader_studies)
-                .exclude(interactive_algorithm="")
-                .values_list("interactive_algorithm", flat=True)
-                .order_by()
-                .distinct()
-            )
-
-        super().save(*args, **kwargs)
-
-        if adding:
-            self.reader_studies.set(reader_studies)
-
-    @property
-    def credits_per_hour(self):
-        if self.interactive_algorithms:
-            return 1000
-        else:
-            return 500
-
-    @property
-    def credits_consumed(self):
-        return ceil(
-            self.duration.total_seconds() / 3600 * self.credits_per_hour
-        )
-
-
-class SessionCostReaderStudy(models.Model):
-    session_cost = models.ForeignKey(SessionCost, on_delete=models.CASCADE)
-    reader_study = models.ForeignKey(
-        "reader_studies.ReaderStudy", on_delete=models.CASCADE
-    )
-
-    class Meta:
-        constraints = [
-            models.UniqueConstraint(
-                fields=["session_cost", "reader_study"],
-                name="unique_session_cost_reader_study",
-            )
-        ]
