+++ app/tests/workstations_tests/factories.py
@@ -1,8 +1,6 @@
-from datetime import timedelta
-
 import factory
 
-from grandchallenge.workstations.models import Feedback, SessionCost
+from grandchallenge.workstations.models import Feedback
 from tests.factories import SessionFactory
 
 
@@ -12,11 +10,3 @@ class FeedbackFactory(factory.django.DjangoModelFactory):
 
     session = factory.SubFactory(SessionFactory)
     user_comment = factory.fuzzy.FuzzyText()
-
-
-class SessionCostFactory(factory.django.DjangoModelFactory):
-    class Meta:
-        model = SessionCost
-
-    session = factory.SubFactory(SessionFactory)
-    duration = timedelta(minutes=15)
